Remove residual fragments from failed video
When a video failed due to unavailable fragments, a subsequent redownload will try to use the residual fragments from the previous download. This is because the download process did not go through to the point to have a video file and a thumbnail downloaded. Since the media.path did not synchronize with the actual calibre-web book path and no timestamp wasn't applied to the media.webpath, these residual fragments affects need to be removed.